third_party/blink/web_tests/platform/generic/http/tests/websocket/unpaired-surrogates-in-message-expected.txt

Checks whether unpaired surrogates are replaced with U+FFFD.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS message is '�'
PASS successfullyParsed is true

TEST COMPLETE
